task_2/fina_llm/requirements_langchain.txt
gradio>=4.0.0
llama-cpp-python>=0.2.0
langchain>=0.0.267
langchain-community>=0.0.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.65.0